+++ src/function/tools/ecef_to_id.rs
@@ -3,6 +3,6 @@ use crate::{
     id::SpaceTimeId,
 };
 
-pub fn ecef_to_id(z: u16, ecef: ECEF) -> SpaceTimeId {
+pub fn ecef_to_id(z: u8, ecef: ECEF) -> SpaceTimeId {
     point_to_id(z, ecef_to_point(ecef))
 }
